+++ mapboard/topology_manager/fixtures/01.1-data-tables.sql
@@ -136,7 +136,10 @@ SELECT ml.id
 FROM {data_schema}.map_layer ml,
      {data_schema}.linework_type lt
 WHERE ml.id = $1.map_layer
-  AND ml.composited_from IS NULL
+  AND NOT EXISTS (
+    SELECT 1 FROM {data_schema}.map_layer_composition m
+    WHERE m.parent_id = ml.id
+  )
   AND lt.id = $1.type
   AND coalesce(lt.topological, true)
   AND ml.topological;